pyproject.toml
[project]
name = "aws-puyo-game"
version = "0.1.0"
description = "AWSのアーキテクチャアイコンを使用したぷよぷよ風ゲーム"
authors = [
    {name = "AWS Game Developer"}
]
requires-python = ">=3.8"
dependencies = [
    "pygame>=2.5.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"